Updating integrated test for RestControllers

+++ src/main/java/org/sanmarcux/samples/sakila/dao/CustomerRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  *
  * @author Cesardl
  */
-public interface CustomerRepository extends JpaRepository<Customer, Short> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 
 }
