+++ src/main/java/org/sanmarcux/samples/sakila/dto/FilmDTO.java
@@ -14,7 +14,7 @@ import java.time.Year;
  */
 public class FilmDTO {
 
-    private Short filmId;
+    private Integer filmId;
     private String title;
     private String description;
     private Year releaseYear;
@@ -32,11 +32,11 @@ public class FilmDTO {
     private Rating rating;
     private String specialFeatures;
 
-    public Short getFilmId() {
+    public Integer getFilmId() {
         return filmId;
     }
 
-    public void setFilmId(Short filmId) {
+    public void setFilmId(Integer filmId) {
         this.filmId = filmId;
     }
