+++ src/main/java/org/sanmarcux/samples/sakila/exceptions/CustomerNotFoundException.java
@@ -2,7 +2,7 @@ package org.sanmarcux.samples.sakila.exceptions;
 
 public class CustomerNotFoundException extends RuntimeException {
 
-    public CustomerNotFoundException(Short id) {
+    public CustomerNotFoundException(Integer id) {
         super("Could not find customer " + id);
     }
 }
